1. Добавление, удаление и изменение товаров. Каждый товар должен иметь изображение. 2. Просмотр деталей товара в отдельной странице. 3. Добавление товаров в корзину и просмотр корзины.

+++ Shop/Data/Models/Car.cs
@@ -9,12 +9,11 @@ namespace Shop.Data.Models
     public class Car
     {
         public int id { get; set; }
-
         public string name { get; set; }
         public string shortDesc { get; set; }
         public string longDesc { get; set; }
         public string  img { get; set; }
-        public ulong price { get; set; }
+        public decimal price { get; set; }
         public bool isFavourite { get; set; }
         public bool iavailable { get; set; }
         public int categoryID { get; set; }
