+++ Shop/Data/Models/OrderDetail.cs
@@ -10,7 +10,7 @@ namespace Shop.Data.Models
         public int id { get; set; }
         public int OrderId { get; set; }
         public int CarId { get; set; }
-        public uint  price { get; set; }
+        public decimal  price { get; set; }
         public virtual Car car { get; set; }
         public virtual Order order  { get; set; }  
     }
